Move progress.ts to /common/vscode/progress.ts
