+++ extensions/ql-vscode/src/common/vscode/progress.ts
