+++ extensions/ql-vscode/src/helpers.ts
@@ -19,7 +19,7 @@ import {
   WorkspaceFolder,
 } from "vscode";
 import { CodeQLCliServer, QlpacksInfo } from "./codeql-cli/cli";
-import { UserCancellationException } from "./progress";
+import { UserCancellationException } from "./common/vscode/progress";
 import { extLogger, OutputChannelLogger } from "./common";
 import { QueryMetadata } from "./pure/interface-types";
 import { telemetryListener } from "./telemetry";
