+++ extensions/ql-vscode/src/language-support/contextual/location-finder.ts
@@ -10,7 +10,7 @@ import {
 } from "../../pure/bqrs-cli-types";
 import { CodeQLCliServer } from "../../codeql-cli/cli";
 import { DatabaseManager, DatabaseItem } from "../../databases/local-databases";
-import { ProgressCallback } from "../../progress";
+import { ProgressCallback } from "../../common/vscode/progress";
 import { KeyType } from "./key-type";
 import {
   qlpackOfDatabase,
