+++ extensions/ql-vscode/src/language-support/contextual/template-provider.ts
@@ -18,7 +18,7 @@ import {
 import { CodeQLCliServer } from "../../codeql-cli/cli";
 import { DatabaseManager } from "../../databases/local-databases";
 import { CachedOperation } from "../../helpers";
-import { ProgressCallback, withProgress } from "../../progress";
+import { ProgressCallback, withProgress } from "../../common/vscode/progress";
 import { KeyType } from "./key-type";
 import {
   FullLocationLink,
