+++ extensions/ql-vscode/src/packaging/packaging.ts
@@ -9,7 +9,7 @@ import {
   ProgressCallback,
   UserCancellationException,
   withProgress,
-} from "../progress";
+} from "../common/vscode/progress";
 import { extLogger } from "../common";
 import { asError, getErrorStack } from "../pure/helpers-pure";
 import { redactableError } from "../pure/errors";
