+++ extensions/ql-vscode/src/query-server/legacy/query-server-client.ts
@@ -13,7 +13,7 @@ import {
   ProgressMessage,
   WithProgressId,
 } from "../../pure/legacy-messages";
-import { ProgressCallback, ProgressTask } from "../../progress";
+import { ProgressCallback, ProgressTask } from "../../common/vscode/progress";
 import { ServerProcess } from "../server-process";
 import { App } from "../../common/app";
