+++ extensions/ql-vscode/src/query-server/legacy/run-queries.ts
@@ -14,7 +14,7 @@ import {
   showAndLogWarningMessage,
   upgradesTmpDir,
 } from "../../helpers";
-import { ProgressCallback } from "../../progress";
+import { ProgressCallback } from "../../common/vscode/progress";
 import { QueryMetadata } from "../../pure/interface-types";
 import { extLogger, Logger } from "../../common";
 import * as messages from "../../pure/legacy-messages";
