+++ extensions/ql-vscode/src/variant-analysis/run-remote-query.ts
@@ -18,7 +18,10 @@ import {
   getRemoteControllerRepo,
   setRemoteControllerRepo,
 } from "../config";
-import { ProgressCallback, UserCancellationException } from "../progress";
+import {
+  ProgressCallback,
+  UserCancellationException,
+} from "../common/vscode/progress";
 import { RequestError } from "@octokit/types/dist-types";
 import { QueryMetadata } from "../pure/interface-types";
 import { getErrorMessage, REPO_REGEX } from "../pure/helpers-pure";
