+++ extensions/ql-vscode/test/vscode-tests/cli-integration/queries.test.ts
@@ -36,7 +36,7 @@ import {
   AppCommandManager,
   QueryServerCommands,
 } from "../../../src/common/commands";
-import { ProgressCallback } from "../../../src/progress";
+import { ProgressCallback } from "../../../src/common/vscode/progress";
 import { withDebugController } from "./debugger/debug-controller";
 
 type DebugMode = "localQueries" | "debug";
