+++ extensions/ql-vscode/test/vscode-tests/global.helper.ts
@@ -14,7 +14,7 @@ import {
 import { CodeQLCliServer } from "../../src/codeql-cli/cli";
 import { removeWorkspaceRefs } from "../../src/variant-analysis/run-remote-query";
 import { CodeQLExtensionInterface } from "../../src/extension";
-import { ProgressCallback } from "../../src/progress";
+import { ProgressCallback } from "../../src/common/vscode/progress";
 import { importArchiveDatabase } from "../../src/databases/database-fetcher";
 import { createMockCommandManager } from "../__mocks__/commandsMock";
