+++ extensions/ql-vscode/test/vscode-tests/minimal-workspace/local-databases.test.ts
@@ -13,7 +13,7 @@ import {
   FullDatabaseOptions,
 } from "../../../src/databases/local-databases";
 import { Logger } from "../../../src/common";
-import { ProgressCallback } from "../../../src/progress";
+import { ProgressCallback } from "../../../src/common/vscode/progress";
 import { CodeQLCliServer, DbInfo } from "../../../src/codeql-cli/cli";
 import {
   encodeArchiveBasePath,
