+++ extensions/ql-vscode/test/vscode-tests/no-workspace/helpers.test.ts
@@ -38,7 +38,7 @@ import {
   showInformationMessageWithAction,
   walkDirectory,
 } from "../../../src/helpers";
-import { reportStreamProgress } from "../../../src/progress";
+import { reportStreamProgress } from "../../../src/common/vscode/progress";
 import { QueryLanguage } from "../../../src/common/query-language";
 import { Setting } from "../../../src/config";
 import { createMockCommandManager } from "../../__mocks__/commandsMock";
